requirement.txt

streamlit
openai-whisper
torch

